Merge two or more color layers based on their transparency values.

Merge several color layers into one based on their transparency and alpha blending. In the final result, the lower layers are visible through the transparent or NA parts (if any) of the upper layers.

Usage

collayers.merge(collayers, opaque_background = "#FFFFFF")

Arguments

collayers

named list, the values must be vectors, matrices or arrays of color strings (as produced by rgb. The names are free form and do not really matter. All values must have the same length.

opaque_background

a single color string or NULL. If a color string, this color will be used as a final opaque background layer to ensure that the returned colors are all opaque. Pass NULL to skip this, which may result in a return value that contains non-opaque color values.

Value

a color layer, i.e., vector of color strings in a hemilist

See also