Unwrap hemi data from a named hemi list.

Unwrap hemi data from a named hemi list.

Usage

hemilist.unwrap(hemi_list, hemi = NULL, allow_null_list = FALSE)

Arguments

hemi_list

named list, can have entries 'lh' and/or 'rh', see hemilist.

hemi

character string, the hemi data name to retrieve from the list. Can be NULL if the list only has a single entry.

allow_null_list

logical, whether to silently return NULL instead of raising an error if 'hemi_list' is NULL

Value

the data

See also