Determine vertex count of right hemi from hemilist of surfaces or the count itself.

Determine vertex count of right hemi from hemilist of surfaces or the count itself.

Usage

numverts.rh(surfaces)

Arguments

surfaces

hemilist of surfaces, or a single integer, which will be interpreted as the number of vertices of the right hemisphere surface.

Value

integer, the number of vertices.