Print a demographics report

Print a demographics report

Usage

report.on.demographics(
  demographics_df,
  group_column_name = NULL,
  paired = FALSE
)

Arguments

demographics_df

a demographics data.frame, as returned by read.md.demographics.

group_column_name,

string or NULL. If given, the column name of the group column. It must be a factor column with 2 levels. Enables group-comparison tests. Defaults to NULL.

paired

Whether the data of the two groups if paired (repeated measurements). Only relevant if group_column_name is given and tests for group differences are included in the report. Defaults to FALSE.

Value

vector of character strings, the lines of the demographics report.

See also